Make IEntityProperties' LangId Lang enum instead of int for easier validation with FluentValidation's IsInEnum.

+++ src/Recipes.Web/Shared/FilterableCardGrid.razor.cs
@@ -18,7 +18,7 @@ public partial class FilterableCardGrid<TItem> where TItem : ComplexEntity
     public string Route { get; set; }
 
     private IOrderedEnumerable<SimpleEntity> entities => Entities
-                                                                .Select(e => e.ToSimpleEntity((int)Lang))
+                                                                .Select(e => e.ToSimpleEntity(Lang))
                                                                 .OrderBy(x => x.Properties.Type)
                                                                 .ThenBy(x => x.Properties.Name);
